文件删除

+++ httpServer/response/response.go
@@ -1,11 +1,11 @@
 package response
 
 type Response struct {
-	Code uint8       `json:"code"`
+	Code int       `json:"code"`
 	Data interface{} `json:"data"`
 }
 
-func GetResponse(code uint8, data interface{}) *Response {
+func GetResponse(code int, data interface{}) *Response {
 	res := new(Response)
 	res.Code, res.Data = code, data
 	return res
